GETP-216 rename: BaseModel, BaseEntity, NotValidDomainModelException을 domain.common으로 이동
