+++ src/main/java/es/princip/getp/domain/common/exception/NotValidDomainModelException.java
@@ -1,4 +1,4 @@
-package es.princip.getp.domain;
+package es.princip.getp.domain.common.exception;
 
 public class NotValidDomainModelException extends RuntimeException {
