+++ src/main/java/es/princip/getp/domain/common/model/Duration.java
@@ -1,6 +1,5 @@
 package es.princip.getp.domain.common.model;
 
-import es.princip.getp.domain.BaseModel;
 import es.princip.getp.domain.common.exception.StartDateIsAfterEndDateException;
 import jakarta.validation.constraints.NotNull;
 import lombok.EqualsAndHashCode;
