+++ src/main/java/es/princip/getp/domain/member/model/Member.java
@@ -1,6 +1,6 @@
 package es.princip.getp.domain.member.model;
 
-import es.princip.getp.domain.BaseEntity;
+import es.princip.getp.domain.common.model.BaseEntity;
 import es.princip.getp.domain.common.model.PhoneNumber;
 import es.princip.getp.domain.member.exception.NotAgreedAllRequiredServiceTermException;
 import es.princip.getp.domain.serviceTerm.model.ServiceTermTag;
