+++ src/main/java/es/princip/getp/domain/people/model/People.java
@@ -1,6 +1,6 @@
 package es.princip.getp.domain.people.model;
 
-import es.princip.getp.domain.BaseEntity;
+import es.princip.getp.domain.common.model.BaseEntity;
 import es.princip.getp.domain.like.command.domain.LikeReceivable;
 import es.princip.getp.domain.like.command.domain.Likeable;
 import es.princip.getp.domain.member.model.Email;
